backend/scripts/requirements.txt
Click==7.0
hjson==3.0.1
Markdown==3.1.1
Jinja2==2.10.3
requests==2.20.0
schedule==0.6.0
humanize==0.5.1
beautifulsoup4==4.8.1
dataclasses==0.7
parsec==3.5
Pillow==6.2.2
PyGithub==1.44.1
python_dateutil==2.8.1
PyYAML==5.2
pretty_errors==1.1.7
